refactor: consolidate DiscoveryArtwork type
