+++ src/Apps/ArtAdvisor/07-Curated-Discovery/Components/Result/ArtworksRail.tsx
@@ -5,23 +5,10 @@ import { BudgetIntent, State } from "Apps/ArtAdvisor/07-Curated-Discovery/App"
 import { Artwork } from "Apps/ArtAdvisor/07-Curated-Discovery/Components/Result/Artwork"
 import { useSystemContext } from "System/Hooks/useSystemContext"
 import { ShelfArtworkPlaceholder } from "Components/Artwork/ShelfArtwork"
+import { DiscoveryArtwork } from "Apps/ArtAdvisor/07-Curated-Discovery/types"
 
 interface ArtworksRailProps {
   state: State
-}
-
-export type DiscoveryArtwork = {
-  id: string
-  internalID: string
-  slug: string
-  title: string
-  date: string
-  rarity: string
-  medium: string
-  materials: string
-  price: string
-  dimensions: string
-  imageUrl: string
 }
 
 function getPriceRange(budgetIntent?: BudgetIntent) {
